bluetooth: Separate bdroid config for bcm4345
Commit 378ffad49c62e4a57e9966dcd880900f32730041 disables BLE_VND_INCLUDED
stating G5 needs it. So apparently it causes problems for other LG devices.

If it doesn't cause an issue with G5, I'm going to assume it's about which
BCM chipset used. Since h850 and rs988 uses bcm4345 (along with h870 and h990),
select a different config for those devices, to enable BLE_VND_INCLUDED.

If this turns out to be wrong, then perhaps the test device was a h830, in
which case the chipset would be bcm4358 (h830, us996, vs995, ls997, h918, h910).